Rename `full_text.py` to `document_full_text.py` for clarity
